updated status effect system removed conditional buff need to clean up files

+++ Assets/Scripts/Abilities/Ab_DamageReduction.cs
@@ -7,7 +7,7 @@ namespace Game.Unit.Ability
     [CreateAssetMenu(menuName = "Game/Abilities/Ab_DamageReduction", fileName = "Ab_DamageReduction")]
     public class Ab_DamageReduction : AbilitySO
     {
-        [SerializeField] private int _damageReduction;
+        [SerializeField] private float _damageReduction;
         
         public override void RegisterTo(UnitObject unit, UnitObject.UnitPartTree.UnitPartTreeNode node)
         {
